# apps/api/openapi/schemawise-api-v1.yaml
openapi: 3.1.0
info:
  title: SchemaWise API
  version: 1.0.0
  description: HTTP contract v1 for synchronous normalization analysis and the public-ready authentication adapter. This describes the HTTP contract, not the engine package version.
tags:
  - { name: Auth, description: Server-side opaque session authentication with session-bound CSRF, credentialed CORS, and auth-specific rate limits. }
  - { name: Computation, description: Public stateless normalization computations. }
paths:
  /api/v1/auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      requestBody: { $ref: '#/components/requestBodies/AuthRequest' }
      responses:
        '201': { $ref: '#/components/responses/AuthCreated' }
        '400': { $ref: '#/components/responses/AuthBadRequest' }
        '403': { $ref: '#/components/responses/AuthForbidden' }
        '409': { $ref: '#/components/responses/AuthConflict' }
        '429': { $ref: '#/components/responses/AuthRateLimited' }
        '500': { $ref: '#/components/responses/AuthInternalError' }
  /api/v1/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      requestBody: { $ref: '#/components/requestBodies/AuthRequest' }
      responses:
        '200': { $ref: '#/components/responses/AuthLoginSuccess' }
        '400': { $ref: '#/components/responses/AuthBadRequest' }
        '401': { $ref: '#/components/responses/AuthUnauthorized' }
        '403': { $ref: '#/components/responses/AuthForbidden' }
        '429': { $ref: '#/components/responses/AuthRateLimited' }
        '500': { $ref: '#/components/responses/AuthInternalError' }
  /api/v1/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      security: [{ sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/AuthSuccess' }
        '401': { $ref: '#/components/responses/AuthUnauthenticated' }
        '500': { $ref: '#/components/responses/AuthInternalError' }
  /api/v1/auth/logout:
    post:
      tags: [Auth]
      operationId: logoutUser
      description: A valid session requires exact browser provenance when supplied and a matching X-CSRF-Token. Missing or inactive sessions remain idempotent.
      security: [{ sessionCookie: [] }, {}]
      parameters: [{ $ref: '#/components/parameters/CsrfHeader' }]
      responses:
        '204': { description: Session revoked when present and cookie cleared idempotently. }
        '403': { $ref: '#/components/responses/AuthForbidden' }
        '500': { $ref: '#/components/responses/AuthInternalError' }
  /api/v1/analysis:
    post:
      tags: [Computation]
      operationId: analyzeSchema
      requestBody: { $ref: '#/components/requestBodies/SchemaInput' }
      responses: { '200': { $ref: '#/components/responses/AnalysisResponse' }, '400': { $ref: '#/components/responses/BadRequest' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/LimitExceeded' }, '500': { $ref: '#/components/responses/InternalError' } }
  /api/v1/closure:
    post:
      tags: [Computation]
      operationId: calculateClosure
      requestBody: { $ref: '#/components/requestBodies/ClosureRequest' }
      responses: { '200': { $ref: '#/components/responses/ClosureResponse' }, '400': { $ref: '#/components/responses/BadRequest' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/LimitExceeded' }, '500': { $ref: '#/components/responses/InternalError' } }
  /api/v1/synthesis/3nf:
    post:
      tags: [Computation]
      operationId: synthesizeThirdNormalForm
      requestBody: { $ref: '#/components/requestBodies/SchemaInput' }
      responses: { '200': { $ref: '#/components/responses/SynthesisResponse' }, '400': { $ref: '#/components/responses/BadRequest' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/LimitExceeded' }, '500': { $ref: '#/components/responses/InternalError' } }
  /api/v1/decomposition/bcnf:
    post:
      tags: [Computation]
      operationId: decomposeBoyceCodd
      requestBody: { $ref: '#/components/requestBodies/SchemaInput' }
      responses: { '200': { $ref: '#/components/responses/BcnfResponse' }, '400': { $ref: '#/components/responses/BadRequest' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/LimitExceeded' }, '500': { $ref: '#/components/responses/InternalError' } }
  /api/v1/analysis/dependency-preservation:
    post:
      tags: [Computation]
      operationId: analyzeDependencyPreservation
      requestBody: { $ref: '#/components/requestBodies/DependencyPreservationRequest' }
      responses: { '200': { $ref: '#/components/responses/DependencyPreservationResponse' }, '400': { $ref: '#/components/responses/BadRequest' }, '413': { $ref: '#/components/responses/PayloadTooLarge' }, '415': { $ref: '#/components/responses/UnsupportedMediaType' }, '422': { $ref: '#/components/responses/LimitExceeded' }, '500': { $ref: '#/components/responses/InternalError' }, 'default': { description: OPERATION_TIMEOUT status is selected by the future execution adapter; no status is frozen in v1., content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } } }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: schemawise_session
      description: Opaque 256-bit session token. The server stores only its SHA-256 digest.
  parameters:
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: false
      description: Required when logout presents an active session; value returned by register, login, or me for that same session.
      schema: { type: string, pattern: '^[A-Za-z0-9_-]{43}$' }
  requestBodies:
    AuthRequest:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/AuthRequestDto' } } }
    SchemaInput:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/SchemaInputDto' } } }
    ClosureRequest:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/ClosureRequestDto' } } }
    DependencyPreservationRequest:
      required: true
      content: { application/json: { schema: { $ref: '#/components/schemas/DependencyPreservationRequestDto' } } }
  responses:
    AuthCreated:
      description: User registered and signed in. The raw session token is returned only in Set-Cookie; the session-bound CSRF token is in the JSON response.
      headers: { Set-Cookie: { schema: { type: string }, description: 'schemawise_session with HttpOnly, Path=/, SameSite=Lax, Max-Age=2592000, Expires, and environment-configured Secure.' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponseDto' } } }
    AuthSuccess:
      description: Authentication succeeded; me returns the stable CSRF token for the active session without replacing its cookie.
      content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponseDto' } } }
    AuthLoginSuccess:
      description: Authentication succeeded with a new independent session and CSRF token. The raw session token is returned only in Set-Cookie.
      headers: { Set-Cookie: { schema: { type: string }, description: 'schemawise_session with HttpOnly, Path=/, SameSite=Lax, Max-Age=2592000, Expires, and environment-configured Secure.' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponseDto' } } }
    AuthBadRequest: { description: Invalid authentication DTO or input policy, content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorEnvelope' } } } }
    AuthUnauthorized: { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorEnvelope' } } } }
    AuthUnauthenticated: { description: Missing or inactive session; an invalid presented cookie is cleared, content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorEnvelope' } } } }
    AuthConflict: { description: Canonical email already exists, content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorEnvelope' } } } }
    AuthForbidden: { description: Browser provenance or session-bound CSRF validation failed with INVALID_CSRF_TOKEN, content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorEnvelope' } } } }
    AuthRateLimited:
      description: An authentication attempt exceeded an applicable IP or email-and-IP limit; the bucket is not disclosed.
      headers: { Retry-After: { schema: { type: integer, minimum: 1 }, description: Seconds until this request can be retried. } }
      content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorEnvelope' } } }
    AuthInternalError: { description: Sanitized authentication infrastructure failure, content: { application/json: { schema: { $ref: '#/components/schemas/AuthErrorEnvelope' } } } }
    AnalysisResponse: { description: Successful analysis, content: { application/json: { schema: { $ref: '#/components/schemas/AnalysisResponseDto' } } } }
    ClosureResponse: { description: Successful closure, content: { application/json: { schema: { $ref: '#/components/schemas/ClosureResponseDto' } } } }
    SynthesisResponse: { description: Successful 3NF synthesis, content: { application/json: { schema: { $ref: '#/components/schemas/ThirdNormalFormSynthesisResponseDto' } } } }
    BcnfResponse: { description: Successful BCNF decomposition, content: { application/json: { schema: { $ref: '#/components/schemas/BcnfDecompositionResponseDto' } } } }
    DependencyPreservationResponse: { description: Successful dependency preservation analysis, content: { application/json: { schema: { $ref: '#/components/schemas/DependencyPreservationResponseDto' } } } }
    BadRequest: { description: Invalid request or scope, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    PayloadTooLarge: { description: Payload exceeds 64 KiB; the future HTTP adapter returns ANALYSIS_LIMIT_EXCEEDED, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    UnsupportedMediaType: { description: Unsupported media type; the future HTTP adapter returns INVALID_REQUEST, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    LimitExceeded: { description: Product limit exceeded (cardinality/length uses 422; payload bytes uses 413 in the future HTTP adapter), content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    InternalError: { description: Internal error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
    AuthRequestDto:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email: { type: string, maxLength: 254 }
        password: { type: string, maxLength: 128, description: Registration requires at least 12 Unicode code points; login permits historical shorter values. }
    PublicUserDto:
      type: object
      required: [id, email]
      additionalProperties: false
      properties:
        id: { type: string, format: uuid }
        email: { type: string, maxLength: 254 }
    AuthResponseDto:
      type: object
      required: [user, csrfToken]
      additionalProperties: false
      properties:
        user: { $ref: '#/components/schemas/PublicUserDto' }
        csrfToken: { type: string, pattern: '^[A-Za-z0-9_-]{43}$', description: HMAC-SHA-256 value bound to the current session; keep in memory and send as X-CSRF-Token on authenticated mutations. }
    AuthErrorCode: { type: string, enum: [INVALID_AUTH_REQUEST, EMAIL_ALREADY_EXISTS, INVALID_CREDENTIALS, UNAUTHENTICATED, INVALID_CSRF_TOKEN, AUTH_RATE_LIMITED, AUTH_INTERNAL_ERROR] }
    AuthErrorEnvelope: { type: object, required: [error], properties: { error: { type: object, required: [code, message], properties: { code: { $ref: '#/components/schemas/AuthErrorCode' }, message: { type: string }, details: { type: object, additionalProperties: true } } } } }
    AttributeDto:
      type: object
      required: [id, name]
      additionalProperties: false
      properties:
        id: { type: string, minLength: 1, maxLength: 64 }
        name: { type: string, minLength: 1, maxLength: 120 }
      description: IDs are case-sensitive and stable per request; uniqueness is validated by Application Layer.
    AttributeSetDto: { type: array, items: { type: string }, uniqueItems: true, description: An attribute ID set; the empty array is valid. }
    RelationDto:
      type: object
      required: [name, attributes]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        attributes: { type: array, minItems: 1, maxItems: 6, items: { $ref: '#/components/schemas/AttributeDto' } }
    FunctionalDependencyDto:
      type: object
      required: [left, right]
      additionalProperties: false
      properties: { left: { $ref: '#/components/schemas/AttributeSetDto' }, right: { $ref: '#/components/schemas/AttributeSetDto' } }
    SchemaInputDto:
      type: object
      required: [relation, functionalDependencies]
      additionalProperties: false
      properties:
        relation: { $ref: '#/components/schemas/RelationDto' }
        functionalDependencies: { type: array, maxItems: 12, items: { $ref: '#/components/schemas/FunctionalDependencyDto' } }
    ClosureRequestDto:
      allOf:
        - { $ref: '#/components/schemas/SchemaInputDto' }
        - type: object
          required: [attributes]
          properties: { attributes: { $ref: '#/components/schemas/AttributeSetDto' } }
    DependencyPreservationRequestDto:
      allOf:
        - { $ref: '#/components/schemas/SchemaInputDto' }
        - type: object
          required: [decomposition]
          properties: { decomposition: { type: array, maxItems: 12, items: { $ref: '#/components/schemas/AttributeSetDto' } } }
    AnalysisResponseDto:
      type: object
      required: [relation, candidateKeys, primeAttributes, minimalCover, normalForms]
      properties:
        relation: { $ref: '#/components/schemas/RelationDto' }
        candidateKeys: { type: array, items: { $ref: '#/components/schemas/AttributeSetDto' } }
        primeAttributes: { $ref: '#/components/schemas/AttributeSetDto' }
        minimalCover: { type: array, items: { $ref: '#/components/schemas/FunctionalDependencyDto' } }
        normalForms:
          type: object
          required: [second, third, bcnf]
          properties:
            second: { $ref: '#/components/schemas/SecondNormalFormDto' }
            third: { $ref: '#/components/schemas/ThirdNormalFormDto' }
            bcnf: { $ref: '#/components/schemas/BcnfNormalFormDto' }
    SecondNormalFormViolationDto:
      type: object
      required: [candidateKey, determinant, dependent]
      properties: { candidateKey: { $ref: '#/components/schemas/AttributeSetDto' }, determinant: { $ref: '#/components/schemas/AttributeSetDto' }, dependent: { type: string } }
    ThirdNormalFormViolationDto:
      type: object
      required: [determinant, dependent]
      properties: { determinant: { $ref: '#/components/schemas/AttributeSetDto' }, dependent: { type: string } }
    SecondNormalFormDto: { type: object, required: [satisfied, violations], properties: { satisfied: { type: boolean }, violations: { type: array, items: { $ref: '#/components/schemas/SecondNormalFormViolationDto' } } } }
    ThirdNormalFormDto: { type: object, required: [satisfied, violations], properties: { satisfied: { type: boolean }, violations: { type: array, items: { $ref: '#/components/schemas/ThirdNormalFormViolationDto' } } } }
    BcnfNormalFormDto: { $ref: '#/components/schemas/ThirdNormalFormDto' }
    ClosureResponseDto: { type: object, required: [closure], properties: { closure: { $ref: '#/components/schemas/AttributeSetDto' } } }
    SynthesizedRelationDto: { type: object, required: [attributes, source], properties: { attributes: { $ref: '#/components/schemas/AttributeSetDto' }, source: { type: string, enum: [minimal-cover, candidate-key] } } }
    ThirdNormalFormSynthesisResponseDto: { type: object, required: [relations, minimalCover, addedCandidateKey], properties: { relations: { type: array, items: { $ref: '#/components/schemas/SynthesizedRelationDto' } }, minimalCover: { type: array, items: { $ref: '#/components/schemas/FunctionalDependencyDto' } }, addedCandidateKey: { oneOf: [{ $ref: '#/components/schemas/AttributeSetDto' }, { type: 'null' }] } } }
    BcnfStepDto: { type: object, required: [source, violation, result], properties: { source: { $ref: '#/components/schemas/AttributeSetDto' }, violation: { $ref: '#/components/schemas/ThirdNormalFormViolationDto' }, result: { type: array, minItems: 2, maxItems: 2, prefixItems: [{ $ref: '#/components/schemas/AttributeSetDto' }, { $ref: '#/components/schemas/AttributeSetDto' }], items: false } } }
    BcnfDecompositionResponseDto: { type: object, required: [relations, steps], properties: { relations: { type: array, items: { type: object, required: [attributes], properties: { attributes: { $ref: '#/components/schemas/AttributeSetDto' } } } }, steps: { type: array, items: { $ref: '#/components/schemas/BcnfStepDto' } } } }
    DependencyPreservationResponseDto: { type: object, required: [preserved, preservedDependencies, lostDependencies], properties: { preserved: { type: boolean }, preservedDependencies: { type: array, items: { $ref: '#/components/schemas/FunctionalDependencyDto' } }, lostDependencies: { type: array, items: { $ref: '#/components/schemas/FunctionalDependencyDto' } } } }
    ErrorCode: { type: string, enum: [INVALID_REQUEST, INVALID_RELATION, INVALID_ATTRIBUTE, INVALID_FUNCTIONAL_DEPENDENCY, ATTRIBUTE_IDENTITY_COLLISION, UNKNOWN_ATTRIBUTE_REFERENCE, SCHEMA_SCOPE_VIOLATION, INCOMPLETE_DECOMPOSITION, ANALYSIS_LIMIT_EXCEEDED, OPERATION_TIMEOUT, INTERNAL_ERROR] }
    ErrorEnvelope: { type: object, required: [error], properties: { error: { type: object, required: [code, message], properties: { code: { $ref: '#/components/schemas/ErrorCode' }, message: { type: string }, details: { type: object, additionalProperties: true } } } } }